Extract `redeem_shares` tests and add missing tests
