+++ zrml/prediction-markets/src/tests/mod.rs
@@ -22,14 +22,15 @@ mod admin_move_market_to_closed;
 mod admin_move_market_to_resolved;
 mod approve_market;
 mod buy_complete_set;
-mod create_market;
 mod close_trusted_market;
+mod create_market;
 mod dispute;
 mod dispute_early_close;
 mod edit_market;
 mod manually_close_market;
 mod on_initialize;
 mod on_market_close;
+mod redeem_shares;
 mod reject_early_close;
 mod reject_market;
 mod report;
